pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "plex_scripts"
version = "1.0"
description = "Various media acquisition/organization automations for plex"
readme = "README.md"

requires-python = ">=3.6"
dependencies = [
    "python-dotenv",
    "plexapi",
    "spotipy",
    "aigpy @ git+https://github.com/Frnot/AIGPY_multiple_artist_tag",
    "tidal-dl",
    "music_tag",
    "billboard.py",
]
